requirements.txt
streamlit
pandas
plotly
Pillow
xlrd>=2.0.1
openpyxl

